reestructuring the app
